New: AbsResponder abstraction New: Monitor abstraction Changed license Respecting standard comment policy
