--- src/main/java/bot/model/query/Regex.java
@@ -1,14 +1,0 @@
-package bot.model.query;
-
-/**
- * Created by Luca Mosetti on 2017
- * <p>
- * REGEXes which Parsers and Builders have to use
- */
-public interface Regex {
-
-    String REGEX = "\n";
-    String TWO_DOTS = ":";
-    String COMMAND = "cmd";
-
-}
