+++ src/main/java/gekoramy/telegram/bot/model/query/Regex.java
@@ -1,0 +1,15 @@
+package gekoramy.telegram.bot.model.query;
+
+/**
+ * REGEXes which Parsers and Builders have to use
+ *
+ * @author Luca Mosetti
+ * @since 2017
+ */
+public interface Regex {
+
+    String REGEX = "\n";
+    String TWO_DOTS = ":";
+    String COMMAND = "cmd";
+
+}
