+++ src/main/java/gekoramy/telegram/bot/responder/type/Handling.java
@@ -1,0 +1,11 @@
+package gekoramy.telegram.bot.responder.type;
+
+/**
+ * @author Luca Mosetti
+ * @since 02/2018
+ */
+public interface Handling {
+
+    void setHandled(boolean handled);
+
+}
