+++ src/main/java/gekoramy/telegram/bot/util/Pair.java
@@ -1,0 +1,46 @@
+package gekoramy.telegram.bot.util;
+
+import java.util.Objects;
+
+/**
+ * @author Luca Mosetti
+ * @since 02/2018
+ */
+public final class Pair<A, B> {
+    private final A a;
+    private final B b;
+
+    private Pair(A a, B b) {
+        this.a = a;
+        this.b = b;
+    }
+
+    public static <A, B> Pair<A, B> of(A a, B b) {
+        return new Pair<>(a, b);
+    }
+
+    public A a() {
+        return a;
+    }
+
+    public B b() {
+        return b;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o)
+            return true;
+
+        if (o == null || getClass() != o.getClass())
+            return false;
+
+        Pair<?, ?> pair = (Pair<?, ?>) o;
+        return Objects.equals(a, pair.a) && Objects.equals(b, pair.b);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(a, b);
+    }
+}
